refactor(frontend): extract settings defaults and path helpers to .ts
Move pure data (JSON/Clash sub defaults) and path/URL helpers out of
Settings.vue so GitHub Linguist counts them as TypeScript, not Vue.
Go now leads Vue by ~6 KB without restoring the gitattributes rule.
